pyproject.toml
[project]
name = "palantyr"
version = "0.1.0"
description = ""
authors = [
    {name = "bm-666",email = "artem.vostrikov.777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiodocker (>=0.24.0,<0.25.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
